kvm: x86/mmu: use a 'unsigned long' view mask from kvm_arch_flush_shadow_all()
This should complete commit f0d3c637889f 'kvm: test_bit() on addresses
not aligned to sizeof(unsigned long) is illegal'.

+++ arch/x86/kvm/x86.c
@@ -10293,7 +10293,7 @@ void kvm_arch_commit_memory_region(struct kvm *kvm,
 
 void kvm_arch_flush_shadow_all(struct kvm *kvm)
 {
-	u16 ept_views_to_keep = 0;
+	unsigned long ept_views_to_keep = 0;
 
 	kvm_mmu_zap_all(kvm, ~ept_views_to_keep);
 }
